requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pocketbase==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
urllib3==2.5.0
